+++ worktile-server/worktile-logic/src/main/java/com/sjtu/worktile/model/TTaskCheckItem.java
@@ -1,11 +1,11 @@
 package com.sjtu.worktile.model;
 
 public class TTaskCheckItem {
-    private Integer id;
+    private Long id;
 
-    private Integer senderId;
+    private Long senderId;
 
-    private Integer taskId;
+    private Long taskId;
 
     private Integer achievement;
 
@@ -17,27 +17,27 @@ public class TTaskCheckItem {
 
     private String content;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getSenderId() {
+    public Long getSenderId() {
         return senderId;
     }
 
-    public void setSenderId(Integer senderId) {
+    public void setSenderId(Long senderId) {
         this.senderId = senderId;
     }
 
-    public Integer getTaskId() {
+    public Long getTaskId() {
         return taskId;
     }
 
-    public void setTaskId(Integer taskId) {
+    public void setTaskId(Long taskId) {
         this.taskId = taskId;
     }
